Fixes for code review. https://github.com/Open-EO/openeo-geotrellis-extensions/issues/265
